requirements.txt
faker
snowflake-connector-python
pandas
pyarrow
great_expectations[snowflake]
sqlalchemy
snowflake-sqlalchemy